Add files via upload
I assumed the role of a Data Scientist / Data Analyst working for a new startup investment firm that helps customers invest their money in stocks. Job is to extract financial data like historical share price and quarterly revenue reportings from various sources using Python libraries and webscraping on popular stocks. After collecting this data I visualized it in a dashboard to identify patterns or trends. The stocks I worked with are Tesla, Amazon, AMD, and GameStop. This is an IBM DATA SCIENCE PROFESSIONAL CERTIFICATE course that I am pursuing and got this project to work on.